Refactor insight creation UI
