+++ client/web/src/enterprise/insights/pages/insights/creation/intro/IntroCreationPage.story.tsx
@@ -6,7 +6,7 @@ import { NOOP_TELEMETRY_SERVICE } from '@sourcegraph/shared/src/telemetry/teleme
 
 import { WebStory } from '../../../../../../components/WebStory'
 import { CodeInsightsBackendContext } from '../../../../core/backend/code-insights-backend-context'
-import { CodeInsightsGqlBackend } from '../../../../core/backend/gql-api/code-insights-gql-backend'
+import { CodeInsightsGqlBackend } from '../../../../core/backend/gql-backend/code-insights-gql-backend'
 
 import { IntroCreationPage } from './IntroCreationPage'
