+++ client/web/src/enterprise/insights/pages/insights/creation/lang-stats/components/lang-stats-insight-creation-content/validators.ts
@@ -3,7 +3,7 @@ import { renderError } from '@sourcegraph/branded/src/components/alerts'
 import { Validator } from '../../../../../../components/form/hooks/useField'
 import { AsyncValidator } from '../../../../../../components/form/hooks/utils/use-async-validation'
 import { createRequiredValidator } from '../../../../../../components/form/validators'
-import { fetchRepositories } from '../../../../../../core/backend/core/requests/fetch-repositories'
+import { fetchRepositories } from '../../../../../../core/backend/gql-backend/methods/get-built-in-insight-data/utils/fetch-repositories'
 
 export const repositoriesFieldValidator: Validator<string> = value => {
     if (value !== undefined && value.trim() === '') {
